db/etl/requirements.txt
pymongo
mysql-connector-python
python-dotenv
pandas
sqlalchemy
pymysql
